backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# API & WebSocket
python-multipart==0.0.6
websockets==12.0
httpx==0.25.2

# Data Processing
pandas==2.1.3
numpy==1.26.2
ta==0.10.2  # Technical Analysis library
scipy==1.11.4

# Data Sources
yfinance==0.2.33
ccxt==4.1.22  # For Binance and other exchanges
requests==2.31.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Caching
redis==5.0.1
aiocache==0.12.2

# Utilities
python-json-logger==2.0.7
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional for ML signals (personal mode)
scikit-learn==1.3.2
joblib==1.3.2